Make support for modifying installations configurable
This patch introduces support for a new setting 'SupportsModify' in the
config.xml file. This is only relevant for Windows installers.

By default (i.e. when omitting) this setting, it
defaults to 'true' which means that the product supports modifying
(i.e. reconfiguring) an existing installations. As such, the 'Modify'
button show in Windows' list of installed programs will be enabled.

Setting this to 'false' will cause the button to be disabled.

Change-Id: I4105f3f0bce67830aa2ee8ae0e6f6abb25c35b30
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>