Fix crash at end of installer on Linux 64bit
The installer occasionally crashed at the very end of the installation,
apparently this is often triggered on Linux 64bit.  It appears that this
is caused by two or more threads using qDebug() concurrently: some
operation still uses OperationTrace::trace() to print status output and
at the same time we try to call showFinishedPage(), which uses qDebug()
to print a not-so-useful message.

Getting qDebug() out of the operations seems to be tough, and defining
QT_NO_DEBUG_OUTPUT at build time to disable *all* qDebug() output feels
like overkill, so let's fix this crash by just not printing a message in
showFinishedPage().

Task-number: QTIFW-905
Change-Id: I4702eb511577148ae6cc0d32f1580d842e58e0b2
Reviewed-by: Katja Marttila <katja.marttila@qt.io>