Fix uninstallation on Windows when target path contains non-ascii characters
On Windows when a target path contains non-ascii characters and MaintenanceTool
has a custom name like Uninstaller it was not possible to start uninstallation
after otherwise successful installation. That was because the renaming of
Uninstaller.exe.new and Uninstaller.dat.new was silently failing somewhere at
the end of installation and there were actaully no files Uninstaller.exe and
Uninstaller.dat.

To fix this encode the vbs file which is passed to cscript as UTF-16 so that it
can handle non-ascii file paths correctly.

Change-Id: Id0feb8f8503594f2a9eedddcaa35ae103ccf2938
Reviewed-by: Karsten Heimrich <karsten.heimrich@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>