Do not fail to install or update if repository is missing
If repository address becomes invalid or is deleted from the server,
update or install fails if installer points to a removed repository.
However, Updates.xml might have RepositoryUpdate with action 'remove'. If
we prevent updating when one of the repositories is not found we never get
to perform the 'remove' action to repository which makes the whole
'remove' action unusable. This commit changes the behavior so that warning
is printed when repository is not found but it does not prevent the
install or update if there are other valid repositories.

Change-Id: Ic75b2e09542e304e09af87c7039798120f58d844
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>