Fix install corrupt with silentInstall
If there are updates available for a component that is executable and the
executable is running, installer is unable to update the component and
results to a corrupted install. Check that there are no processes running
from the install tree and cancel update if there are any. The update is
canceled if there are processes running even if the process does not get
update.

Task-number: QTIFW-969
Change-Id: I5614cdd21d8538ccf8b0d114d3616804fe367667
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>