revamp build of translations
instead of using a hand-crafted qrc file, auto-generate one. this allows
us to collect available translations dynamically, which means that we
don't need to limit ourselves to the state of qt 5.5.

as a side effect, this removes the possibility to compile the qt
translations ourselves. i'm not sure what the point of that was, as
every qt installation would have pre-compiled qm files.

Change-Id: Ie88327e848c2498eb2a63a895a80e1cd603619b8
Reviewed-by: Katja Marttila <katja.marttila@qt.io>