Add possibility to reopen admin query
In windows if admin rights are asked, you have two minutes to answer
to the query. After that installing is not possible as there is no way to
go back to admin query.
Adding a possibility to restart the admin right query when 'No' pressed
or two minutes timeframe has passed in Windows. Requires
DisableAuthorizationFallback set to true, otherwize user is asked to use
fallback '--startserver'.

Task-number: QTIFW-988
Change-Id: Ide33b588299c981cfe653e7f53e7316a768d87bb
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>