Docs: Remove dependency to Qt's macros.qdocconf
The macros.qdocconf from Qt's global documentation config includes a
macro that uses the value of QT_VERSION, which is not defined for
IFW.

The documentation build therefore fails with
'error: Environment variable 'QT_VERSION' undefined'.

To work around this, load the macros we use from a local qdocconf
file.

Change-Id: I1dc02dffc758af8e855e42a0fceabb84434876d2
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>