Allow maintenancetool signing in Windows
Maintenancetool signing was broke as installer modified the
maintenancetool binary. Fixed so that maintenancetool is not
modified, instead the needed data is written to a separate
installer.dat file. Installer.dat is written and needed only after
install so we can continue using and deploying only one executable.

Task-number: QTIFW-667
Change-Id: I30bf2ebe81d7c7146a78840d234c5c813f8e1da5
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>