Optimize ExctractArchiveOperation::fileFinished()
This code was spending a lot of time detaching and
copying the file list, due to exporting it via the
setValue() call on each file finished iteration. This
was causing the install process to slow down, and
also made the UI freeze, giving the impression that the
install process had stopped.

Add the file paths to a local list instead, and then
call setValue when file extraction is complete.

Profiling shows that the installer now spends 95% of its
time extracting files on the ExctractArchiveOperation,
thread and 5% on the main thread.

Fixes: QTBUG-51337
Change-Id: Ieb4c44c03cb28eb46928730ca023f7419d72e45b
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: David Faure <david.faure@kdab.com>