Remove 'Your install seems to be corrupted' messagebox
Messagebox is replaced with log. The critical messagebox is a bit harsh
as you can continue using the installer despite of the 'corrupted
install' message.
'Your install seems to be corrupted' can happen for number of reasons:
the installer for some reason was not able to write the needed information
to components.xml or maintenancetool.dat. Reason could be that the
installer was forcely interrupted or crashed during write, or the files
cannot be opened for write. 'Your install seems to be corrupted'
occurs if components.xml contains an element which has no operations in
maintenancetool.dat or vice versa. If you get an error
"Critical: Operations missing from installed packages
("missing.package.name"), remove the package from
components.xml to get rid of the log. Error: "Critical: Orphaned
operations("package.name")" you can ignore
Or you can install the package again if it is still found from
repositories.

Task-number: QTIFW-1003
Change-Id: I02c0a945ca3ce30ef4b0a80190d91638ab54c6bf
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>