Replace usage of derived LazyPlainTextEdit class with QTextEdit
Remove LazyPlainTextEdit class and replace usage in
PerformInstallationForm details browser to QTextEdit. This adds
support for RTL text alignment as the base class of the removed
class, QPlainTextEdit does not fully support RTL and automatic
alignment. Text will now be automatically aligned based on the
direction of the characters.

This also improves the usability of the details browser, as it
does not force scrolling to bottom if the user has scrolled the
widget contents to inspect previous lines.

Task-number: QTIFW-1994
Change-Id: If01eaea121000edc0b5039edfede88ffb7bacd6f
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>