Fix 'Bad allocation' error in Windows
Due to bug in QTextDocument (QTBUG-65865), installer will crash if it
will write a lot of logs to 'Details View' in 'PerformInstallationPage'.
Fixed so that instead of writing all extracted filenames to log, write
only the name of package which is extracted. This will reduce the huge
amount of log written and thus prevents the 'Bad allocation' crash.

Task-number: QTIFW-1242
Change-Id: I98ff0df25bcc10d0f7bb79d6c4010665b5e2ac51
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>