Allow unstable components when categories are used
Unstable component means that the component has unresolved dependency,
script error etc. Installer cannot recover if we are already in
component selection view and fetch new packages from a category which
has unstable components if AllowUnstableComponent is false. Fixed so
that allowing unstable components is forced when categories are used.
Ideal solution would be to recover and rollback the install tree if
unstable components are found but that requires huge changes to IFW.

Task-id: QTIFW-1257
Change-Id: I786df1b8b54c238f50e15b94a06005e244417c97
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>