Fix essential update
Essential components should be updated before any other component is
updated/installed. This did not happen always as when one essential
component was found and that did not have any updates, the search was
stopped and normal install flow was continued. Fixed so that all
essential components are checked if they contain update.

Task-number: QTIFW-1299
Change-Id: I754c50f672dd5f13105c710522603e90799d61c5
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>