Hide scrollarea background while fetching categories
Previously we only hide the label inside the scroll area while fetching
categories, which works if the scrollarea has no background. In macOS
there is a background color in the scrollarea by default so hide the
scrollarea instead of the label to prevent the scrollarea colored
background to be visible.

Change-Id: I2ddc97f7ae9a268b8972c91ab6519381fe7b689c
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>