Add early prompt to run MaintenanceTool as admin when needed
Currently IFW does not support manipulating Windows Access Control Lists
for files. If a user does not have reading permissions to installer.dat
it will throw error on QInstaller::openForRead() before the wizard GUI
is shown. At very least, add a MessageBox to prompt the user to run the
executable again as administrator in such a case.

Task-number: QTIFW-1392
Change-Id: I9e20c34b215ee697118af9c15a7e344f9cfc3c3f
Reviewed-by: Katja Marttila <katja.marttila@qt.io>