Enable QInstaller::setDefaultFilePermissions() also on Windows
In some cases, for example in CreateLocalRepository files may be left
with legacy read-only flag on Windows. Even though the permission
entries in its Access Control List are sane, this will block all writing
operations to the file.

Enable setDefaultFilePermissions() on Windows as this will unset the
read-only flag if it is set. This still does not affect ACLs.

Task-number: QTIFW-1464
Change-Id: I95399e7cb83cb136ef1c978a05f79b6a6790b7e4
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>