Fix progressbar length in component selection page
Also use QStackedLayout to transform between treeview and progressbar
views. This simplifies the code as we hide the treeview, categories etc.
and show progressbar instead when we are fetching a new category.
QStackedLayout makes it easier as we then don't have to hide/show
individual widgets we want to show.

Change-Id: I405030e9629052c5ca773a91bc44b66d3d26e137
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>