Fix EnvironmentVariableOp undo behavior with non-persistent variables
The return value for undoOperation() was incorrect when dealing with
non-persistent variable, causing an empty installer error dialog to
appear if the current env does not contain exactly same variable value
as on installation.

The return value inside !isPersistent condition scope should always be
true regardless of if the temporary variable set on installation still
exists on current environment or contains the same value.

Change-Id: If7deaa310d73a371194138fa1e2c382614452e26
Reviewed-by: Katja Marttila <katja.marttila@qt.io>