Provide a means to use fixed translations for Qt and IFW
Since building Qt statically and then IFW can be seen as a lot to
do just to be able to provide a fixed translation for a Qt or IFW
string then it is better to provide a means to override these
translations.

So if a translation is found inside :/translations_new then it will
use this one instead of the one found at :/translations for Qt or
IFW strings. This can be provided via the -r option when using
binarycreator then.

Change-Id: I007c45ff96ccf5d73b54a5df8a80a2a05c2f35c6
Reviewed-by: Katja Marttila <katja.marttila@qt.io>