Change IntroductionPage radio button handling in maintenance tool
If magic binary marker is set to either updater or package manager,
act as before.

If the marker is set to uninstaller (default for maintenance tool),
decide the default checked button based on if any valid repositories
are available - a prerequisite for package manager and updater
actions to succeed.

If repositories are available the default is "Add or remove components",
otherwise "Remove all components".

Modify validRepositoriesAvailable() implementation to suit this change
and better match its function name.

Task-number: QTIFW-349
Change-Id: Idf92ad3a2b294cc1ecb939d716f2852ad1872f15
Reviewed-by: Katja Marttila <katja.marttila@qt.io>