Add function to allow additional processes to be run
Maintenancetool does not allow updating install if running processes are
found from install directory. Only maintenancetool process was allowed,
otherwise the install failed with query to shut down processes. This
commit adds a function to extend the allowed processes to be
run, this is mandatory when running tests as test executable lies in the
install directory.

Task-number: QTIFW-1019
Change-Id: I5422b0327e3f77edf3f9c88c74f40eff5cc8c830
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>