Answer queries from command line
Standard input is possible in Windows only in console application. This
change makes installers to be console applications by default. If
started as gui from command line, console is released and application
is restarted so cmd.exe won't be waiting for application to finish.
If started from explorer, console is launched but released immediately.
This may be seen on screen with short console flash.

Using command line interface messagebox strings and possible answer
values for it are printed to console. Answers must be given in case
sensitive format as StandardButton QMetaEnum is utilized when reading
button values. This ensures that if StandardButtons are
added/removed/changed the code continues to work.

Also added a new option 'default-answer' which will answer the questions
with default button values without user interaction.

Task-number: QTIFW-1739
Change-Id: I710f921f5c927dbd13090af32f93656dded84e72
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>