Fix InstallAction type when updating components
Components should be selected for update using model, not components own
check state so model values stay in sync. Wrong InstallAction type will
cause component to think it will be installed when that is not the case.
This caused problems in install scripts when trying to figure out if
component is going to be installed or not.

Task-number: QTIFW-1830
Change-Id: I367f476de96f27b27d74e06a6fb630f0876924d2
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>