Fix styling of QMessageBoxes
MessageBoxHandler will decide an appropriate parent widget for new
QMessageBox objects just before they are shown. This is not quaranteed
to be a PackageManagerGui instance, for example when the installer
wizard does not have focus, which means the style properties do not
get inherited.

Rather set the used style sheet for the whole application using
QApplication::setStyleSheet().

Task-number: QTIFW-2121
Change-Id: I6f2131279d6f6d3dd41b9eacab8404770aad09a3
Reviewed-by: Katja Marttila <katja.marttila@qt.io>