Reorder ComponentSelectionPage widgets
Move the category widget from the leftmost column in the page
as it takes a lot of horizontal space, and we don't utilize the
empty space below it.

Add a QToolBox widget to the rightmost QGridLayout column to
display either component information or repository categories
at a time. If category widget is not shown, display the description
widgets directly on the page's QGridLayout without the tool box.

Task-number: QTIFW-2096
Change-Id: I04266789e3e4b7ce352b7e0a7230ab80c5e295ce
Reviewed-by: Katja Marttila <katja.marttila@qt.io>