Load and install translations as early as possible
This will ensure that any early error messages are translated if
there are any translations for them. Such as the already running
instance string. As a result this means that a check if the
installer is already running is done slightly later to allow the
loading of the translations first before checking.

Fixes: QTIFW-1888
Change-Id: I75ef22d98ca4de9a05e7dd485b5afed5652529bc
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>