Fix hang when canceling metadatajob by an external call
Installer may leave an event loop running in DownloadFileTask::doTask
when shutting down, if Downloader::testCanceled() was not yet triggered
by connection events, or by a timer timeout used to check if the user
has canceled download while the connection is blocked.

Fix by waiting for the tasks to finish when canceled.

Task-number: QTIFW-2282
Change-Id: I8800cdb3c368da3edaf8def50e3b8e837d3e993c
Reviewed-by: Katja Marttila <katja.marttila@qt.io>