Add possibility to search for components in ComponentSelectionPage
Introduce ComponentSortFilterProxyModel, that compared to its base
class QSortFilterProxyModel, accepts also child indexes of a source
model index which is directly accepted by the filter.

Also remove emitting of QAbstractItemModel::dataChanged() for each
item in the ComponentModel, which was done regardless of if data at
index was changed or not, after reset or when check state or data for
any item was changed. If the ComponentModel acts as a source model
for a proxy model with 'dynamicSortFilter' property set to true, the
extra emits hurt the performance as the model gets filtered again for
each source model change. Actual changes done to the ComponentModel
are still notified.

Task-number: QTIFW-1404
Change-Id: Iea696f8b9abf79d877fb2568488b09dc3cb061be
Reviewed-by: Katja Marttila <katja.marttila@qt.io>