LibArchiveWrapper: use local zero-timer for processing server signals
Now that the server process may request file seeks in addition to reads
for archive formats supporting them, the interval which we process the
requests will have a big impact on the extracting speed.

Also shrink the read buffer as its size isn't as significant for the
performance after this change.

Task-number: QTIFW-2384
Change-Id: Iaf296833c6ff1acfc666a6e82f661e1a40c77b61
Reviewed-by: Katja Marttila <katja.marttila@qt.io>