RemoteObject: fix RemoteServer not processing the "Destroy" command
The RemoteObject destructor did not wait that the actual command data
was written to the socket before it was closed.

The incorrect behavior could be noticed with the recently introduced
LibArchiveWrapperPrivate subclass of RemoteObject. Deleting the client
side object would leave the associated RemoteServerConnection thread
owning the wrapped object occasionally in a running state.

Change-Id: I5297fae15d21363227f63b5340946c03cfd03400
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>
Reviewed-by: Katja Marttila <katja.marttila@qt.io>