Add support for moving child components with <TreeName>
Add new optional "moveChildren" attribute to <TreeName> element.
Children of components declaring "moveChildren=true" attribute are
calculated a value for an automatic tree name element.

Multiple components in a single tree branch can declare a
tree name, the order which the relocation happens is from leaf
components to root components.

Components may be moved under another tree name target.

Components may be moved to an existing identifier part that
does not have a component, for example in the following
repository structure:

    rootA.childA
    rootB
    rootB.childB

the "rootB" component can legally declare a "rootA" tree name.

The tree names of components become static after installed. If
a repository declares a new tree name for a component that is
installed, it is only applied after updating in maintenance tool.
Child components, which have an automatic tree name, are moved if
the parent component is updated with a new tree name, however.

Task-number: QTIFW-2380
Change-Id: I9c44a114d3c1248b7e2dd4f0b5cda0739af102f3
Reviewed-by: Katja Marttila <katja.marttila@qt.io>