Fix "Empty filename passed to function" warning on target directory page
When verifying the selected target directory for forbidden directories
(user home and root), we incorrectly checked if the directory was
empty regardless of if the path of the QDir object was empty.

No need to check this as installing into either of these should be
prevented - the folder may be populated after initial installation and
installer would then nuke it on uninstallation.

Also fix related warning about deprecated '=' operator.

Task-number: QTIFW-2424
Change-Id: If4835dcc719e09e0a51a13149ae05d75c0e2b359
Reviewed-by: Katja Marttila <katja.marttila@qt.io>