Fix selecting target directory when elevated server process is active
There is a disparency in the return value of QFileInfo::isFile() on
Windows whether QFileInfo uses QFileSystemEngine directly (for local
files, returns false) or with another file engine (QFSFileEngine in this
case, returns true), when the file does not exist on disk.

Installer registers its own file engine handler that handles file
engine creation. It creates custom RemoteFileEngine instances suitable
for writing into more restrictive locations when the connection to the
elevated installer server process is active.

When calling RemoteFileEngine::fileFlags() with TypesMask parameter, the
returned combination of enum values includes FileFlag::FileType when
the given file does not exist, this can be reproduced with both the
active remote server connection that forwards the file engine calls to
the server process, and when the composed QFSFileEngine object is used
directly without the remote connection.

Fix by adding a check if the given target directory path exists, if it
doesn't we can assume it is a directory to be created by the installer.
Also rename function to better match what it does.

Task-number: QTIFW-2383
Change-Id: I906b0676802d9d94e70abeb2dea1f71866cbbec8
Reviewed-by: Katja Marttila <katja.marttila@qt.io>