Tests: suppress "QTextStream: No device" warnings
QInstaller::init() registers the custom message handler used by
installer, which formats and prints the messages and writes the
installation log. We clear the buffer device used for the log contents
before it gets written on disk when destructing the installer's
PackageManagerCore object, but on tests there might be several core
objects.

Use a silent message handler instead.

Change-Id: I6694bca1dfffaffd1389f8f619531f1b0052c090
Reviewed-by: Katja Marttila <katja.marttila@qt.io>
Reviewed-by: Iikka Eklund <iikka.eklund@qt.io>