MT: Fix setting incorrect TargetDir value when started as root
Maintenance tool was using the vendor configured default target
directory for system wide installations instead of the actual
application path.

In addition to possibly several other things breaking as the variable
is used extensively by the framework, this would make the installation
appear corrupted as the local package hub file could not be read from
the incorrect location.

Task-number: QTIFW-2486
Change-Id: I31676e57d83409022e6908e64fac636b12d8262d
Reviewed-by: Katja Marttila <katja.marttila@qt.io>