Use of separators in Execute operation
Execute -operation might be strict with the path separators. If wrong
separators are used the Execute -operation might fail. Fixed so that
Execute -operations paths are saved to dat file without converting to
normalized separators. As the target dir is saved as @RELOCATABLE_PATH@
also for Execute -operation, check from the next available separator
which separator should be used for operation.

Task-number: QTIFW-2501
Change-Id: Ia9969bbe3177a6b2daab21c026df88d4e2ec2bb0
Reviewed-by: Arttu Tarkiainen <arttu.tarkiainen@qt.io>